requirements/kozmic.txt
PyMySQL==0.6.1
SQLAlchemy==0.8.3
redis==2.8.0
celery==3.1.4

Flask==0.10.1
Jinja2==2.7.1
Werkzeug==0.9.4
WTForms==1.0.5
Flask-SQLAlchemy==1.0
Flask-Script==0.6.2
Flask-WTF==0.9.3
Flask-Login==0.2.7
Flask-Principal==0.4.0
Flask-Migrate==0.1.4
Flask-OAuth==0.12
Flask-Assets==0.8

pytailf==1.1
pycrypto==2.6.1
furl==0.3.5
github3.py==0.7.1
certifi==0.0.8
ipython==1.1.0  # to have nice ./manage.py shell
docker-py==0.2.2
requests==1.2.0  # due to https://github.com/dotcloud/docker-py/issues/47
raven==3.5.2
